Fixed empty interface list after extension reload (e.g. after suspend).
Occasionally it takes some time (10-20s) for the network interface on my system
to go up after resuming from suspend. This meant the if list was empty,
which caused system-monitor not to update the network metrics. This fix keeps
calling update_iface_list as long as if is empty.

Note that on my systemd-network system the if list is populated by parsing
/proc/net/dev, so I moved the logic to the new update_iface_list_proc() method.